Fixed update bug
Updating models wouldn't work due to the fact that dates would come
in as Strings via REST and the app failed to correctly convert them
to dates for bookshelf. Also, another issue was, that the created_at
and updated_at timestamps were also written to the database on each
update, the timestamps are now removed from the model before saving.